### Added a Markdown & HTML conversion
Added a conversion to Markdown or HTML for the matrices. The way it's done is by giving a name to each node, and then generating a simple Markdown/HTML table. This required to add a optional names to nodes, which led to updates to things related to this (ex. in the display of the graph, or in the return of the clusterMatrice() method)