### Added a Matrice/Graph view selector
Added a way to select which Matrice/Graph should be displayed. This can either be done with the "View" menu, or by using Maj+[1-9] (for Matrices) or Ctrl+[1-9] (for Graphs).

Alongside this, added a widget to display Matrices, renamed the Graph class to GraphWidget, updated some methods (mainly returning `Matrice* `instead of `Matrice`, and added const to some methods), and added a template for import/export of files. Also, commented the tests done at the start in the console, as it's not as important with the new way to visualize Matrices/Graphs.

> At this point, the code probably needs some cleanup, which should be done in the next commit.