### Added examples
Added examples for sizes 3, 4, (6), 8, 10, 12, 15, (20*), 30, 50. Values in parenthesis already existed, and matrices for size 20 were completed based on the existing one. Each size contains a matrice with weights 0 or 1, a matrice with random weights (but the same shape as the previous one) and a matrice for installation costs

Also python scripts were added to create those matrices

> Note that these scripts were cleaned up **after** the initial generation, and weren't tested since (except for install_costs), so they might not work perfectly